ctp2_code\robot\pathing\CityAstar.cpp: corrected endless path computation
Rest: added casts, explicit types, cleaned up (no functional changes)


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@510 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/robot/aibackdoor/bit_table.h
@@ -17,8 +17,6 @@
 //
 // Compiler flags
 // 
-// _MSC_VER		
-// - Compiler version (for the Microsoft C++ compiler only)
 //
 //----------------------------------------------------------------------------
 //
@@ -86,7 +84,7 @@ public:
 		if (rhs.m_data)
         {
             m_data = new sint32[m_total_len];
-            std::copy(rhs.m_data, rhs.m_data + m_total_len, m_data);
+            std::copy(rhs.m_data, rhs.m_data + static_cast<size_t>(m_total_len), m_data);
         }
 	};
